move function tests to test_iam.py
