+++ pvlib/test/test_pvsystem.py
@@ -82,27 +82,6 @@ def test_systemdef_dict():
     assert expected == pvsystem.systemdef(meta, 5, 0, .1, 5, 5)
 
 
-@needs_numpy_1_10
-def test_ashraeiam():
-    thetas = np.array([-90. , -67.5, -45. , -22.5,   0. ,  22.5,  45. ,  67.5, 89.,  90. , np.nan])
-    iam = pvsystem.ashraeiam(thetas, .05)
-    expected = np.array([        0,  0.9193437 ,  0.97928932,  0.99588039,  1.        ,
-        0.99588039,  0.97928932,  0.9193437 ,         0, 0,  np.nan])
-    assert_allclose(iam, expected, equal_nan=True)
-
-
-@needs_numpy_1_10
-def test_ashraeiam_scalar():
-    thetas = -45.
-    iam = pvsystem.ashraeiam(thetas, .05)
-    expected = 0.97928932
-    assert_allclose(iam, expected, equal_nan=True)
-    thetas = np.nan
-    iam = pvsystem.ashraeiam(thetas, .05)
-    expected = np.nan
-    assert_allclose(iam, expected, equal_nan=True)
-
-
 def test_PVSystem_ashraeiam(mocker):
     mocker.spy(pvsystem, 'ashraeiam')
     module_parameters = pd.Series({'b': 0.05})
@@ -113,33 +92,6 @@ def test_PVSystem_ashraeiam(mocker):
     assert iam < 1.
 
 
-@needs_numpy_1_10
-def test_physicaliam():
-    aoi = np.array([-90. , -67.5, -45. , -22.5,   0. ,  22.5,  45. ,  67.5,  90. , np.nan])
-    iam = pvsystem.physicaliam(aoi, 1.526, 0.002, 4)
-    expected = np.array([        0,  0.8893998,  0.98797788,  0.99926198,         1,
-        0.99926198,  0.98797788,  0.8893998,         0, np.nan])
-    assert_allclose(iam, expected, equal_nan=True)
-
-    # GitHub issue 397
-    aoi = pd.Series(aoi)
-    iam = pvsystem.physicaliam(aoi, 1.526, 0.002, 4)
-    expected = pd.Series(expected)
-    assert_series_equal(iam, expected)
-
-
-@needs_numpy_1_10
-def test_physicaliam_scalar():
-    aoi = -45.
-    iam = pvsystem.physicaliam(aoi, 1.526, 0.002, 4)
-    expected = 0.98797788
-    assert_allclose(iam, expected, equal_nan=True)
-    aoi = np.nan
-    iam = pvsystem.physicaliam(aoi, 1.526, 0.002, 4)
-    expected = np.nan
-    assert_allclose(iam, expected, equal_nan=True)
-
-
 def test_PVSystem_physicaliam(mocker):
     module_parameters = pd.Series({'K': 4, 'L': 0.002, 'n': 1.526})
     system = pvsystem.PVSystem(module_parameters=module_parameters)
@@ -148,88 +100,6 @@ def test_PVSystem_physicaliam(mocker):
     iam = system.physicaliam(thetas)
     pvsystem.physicaliam.assert_called_once_with(thetas, **module_parameters)
     assert iam < 1.
-
-
-def test_iam_martin_ruiz():
-
-    aoi = 45.
-    a_r = 0.16
-    expected = 0.98986965
-
-    # will fail of default values change
-    iam = pvsystem.iam_martin_ruiz(aoi)
-    assert_allclose(iam, expected)
-    # will fail of parameter names change
-    iam = pvsystem.iam_martin_ruiz(aoi=aoi, a_r=a_r)
-    assert_allclose(iam, expected)
-
-    a_r = 0.18
-    aoi = [-100, -60, 0, 60, 100, np.nan, np.inf]
-    expected = [0.0, 0.9414631, 1.0, 0.9414631, 0.0, np.nan, 0.0]
-
-    # check out of range of inputs as list
-    iam = pvsystem.iam_martin_ruiz(aoi, a_r)
-    assert_allclose(iam, expected, equal_nan=True)
-
-    # check out of range of inputs as array
-    iam = pvsystem.iam_martin_ruiz(np.array(aoi), a_r)
-    assert_allclose(iam, expected, equal_nan=True)
-
-    # check out of range of inputs as Series
-    aoi = pd.Series(aoi)
-    expected = pd.Series(expected)
-    iam = pvsystem.iam_martin_ruiz(aoi, a_r)
-    assert_series_equal(iam, expected)
-
-    # check exception clause
-    with pytest.raises(RuntimeError):
-        pvsystem.iam_martin_ruiz(0.0, a_r=0.0)
-
-
-@requires_scipy
-def test_iam_interp():
-
-    aoi_meas = [0.0, 45.0, 65.0, 75.0]
-    iam_meas = [1.0,  0.9,  0.8,  0.6]
-
-    # simple default linear method
-    aoi = 55.0
-    expected = 0.85
-    iam = pvsystem.iam_interp(aoi, aoi_meas, iam_meas)
-    assert_allclose(iam, expected)
-
-    # simple non-default method
-    aoi = 55.0
-    expected = 0.8878062
-    iam = pvsystem.iam_interp(aoi, aoi_meas, iam_meas, method='cubic')
-    assert_allclose(iam, expected)
-
-    # check with all reference values
-    aoi = aoi_meas
-    expected = iam_meas
-    iam = pvsystem.iam_interp(aoi, aoi_meas, iam_meas)
-    assert_allclose(iam, expected)
-
-    # check normalization and Series
-    aoi = pd.Series(aoi)
-    expected = pd.Series(expected)
-    iam_mult = np.multiply(0.9, iam_meas)
-    iam = pvsystem.iam_interp(aoi, aoi_meas, iam_mult, normalize=True)
-    assert_series_equal(iam, expected)
-
-    # check beyond reference values
-    aoi = [-45, 0, 45, 85, 90, 95, 100, 105, 110]
-    expected = [0.9, 1.0, 0.9, 0.4, 0.3, 0.2, 0.1, 0.0, 0.0]
-    iam = pvsystem.iam_interp(aoi, aoi_meas, iam_meas)
-    assert_allclose(iam, expected)
-
-    # check exception clause
-    with pytest.raises(ValueError):
-        pvsystem.iam_interp(0.0, [0], [1])
-
-    # check exception clause
-    with pytest.raises(ValueError):
-        pvsystem.iam_interp(0.0, [0, 90], [1, -1])
 
 
 @pytest.fixture(scope="session")
